Refatora funcionalidade de importação de CSV
