+++ backend/spec/requests/import_csv_spec.rb
@@ -12,6 +12,8 @@ describe "POST '/import'" do
     expect(Test.all.count).to eq 4
     expect(last_response.status).to eq 200
     expect(last_response.content_type).to include 'application/json'
+    json_response = JSON.parse(last_response.body, symbolize_names: true)
+    expect(json_response[:message]).to eq 'Data imported successfully.'
   end
 
   it 'returns an error if file type is not supported' do
